requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.1
tiktoken==0.5.1
numpy==1.24.3
python-multipart==0.0.6